requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
shapely==2.0.2
python-dotenv==1.0.0
websockets==12.0
redis==5.0.1
psutil==5.9.6
numpy==1.26.4
requests==2.31.0
python-multipart==0.0.6 